move http_service to tools
